Extract runtime issue integration tests
