+++ docs/DECISIONS.md
@@ -263,3 +263,4 @@
 - Workspace review integration coverage lives in `WorkspaceReviewIntegrationTests`, not `WorkspaceModelTests`. Apply-patch review diff refresh, local and SSH Remote stage/restore actions, hunk staging, review comments, tool cards, and review-surface visibility cross the workspace model, git tools, remote SSH executor, transcript events, and review surface; keep those flows together while pure review comment planning and git command construction stay in focused unit tests.
 - Composer integration coverage lives in `WorkspaceComposerIntegrationTests`, not `WorkspaceModelTests`. Composer submit, tool-card creation, artifact surfacing, Computer Use dispatch, queued-tool progress, streaming assistant drafts, cancellation notices, empty-draft no-ops, and selection races cross the workspace model, agent runner, transcript events, tool cards, and top-bar status; keep those flows together while pure submission/cancellation planning stays in focused planner tests.
 - Message feedback and artifact preview coverage should stay out of `WorkspaceModelTests`. Feedback persistence lives in `WorkspaceFeedbackIntegrationTests` because it crosses model mutation, thread events, and transcript surfaces. Pure `ToolArtifactState` image/document/text preview derivation lives in `QuillCodeToolCardSurfaceTests` because it is value-type surface behavior independent of workspace state.
+- Runtime issue integration coverage lives in `WorkspaceRuntimeIssueIntegrationTests`, not `WorkspaceModelTests`. Applying runtime status, projecting runtime issues into top-bar/settings surfaces, diagnostic redaction, and retry-last-turn mutation cross runtime state, workspace surfaces, commands, and composer state; keep pure issue construction in `WorkspaceRuntimeIssueBuilderTests` and recovery-action mapping in `QuillCodeRuntimeIssueRecoveryPlannerTests`.
